Crate rbot_discord

discord-db is a Rust Discord BOT.

Re-exports

extern crate std;

extern crate chrono;

extern crate imap;

extern crate native_tls;

extern crate rand;

extern crate reqwest;

extern crate serde;

extern crate serde_json;

extern crate serenity;

extern crate time;

extern crate lazy_static;

extern crate serde_derive;

use std::prelude::v1::*;

use core::files;

use features::mail::Info;

use features::notify;

use serenity::model::id::ChannelId;

use std::io;

use std::sync::Arc;

use std::sync::RwLock;

use chrono::prelude::*;

Modules

core
The base of the program containing the abstractions for files and connection to discord.

features
Features for the bot

Structs

CREDENTIALS_FILE
Represent the file that store the credential

Credentials
Store the credential for the mail account, and discord token.

MAIL_INFO_FILE
Represent the file for mails config

MAIL_LOCK
Is the last mail text, stored to be use on call back

NOTIFY_EVENT_FILE

Constants

_IMPL_DESERIALIZE_FOR_Credentials

_IMPL_SERIALIZE_FOR_Credentials

Statics

CREDENTIALS_FILE

MAIL_INFO_FILE

MAIL_LOCK

NOTIFY_EVENT_FILE

Functions

main
We run the core and we loop on a basic cmd.