Module rbot_discord::core::files

Abstraction of a file handler, allow you to open a file and give it a structure, to serialize/deserialize from.

Re-exports

use serde;

use serde_json;

use serde_json::from_reader;

use serde_json::from_str;

use serde_json::to_string;

use serde_json::Value;

use std::fs::File;

use std::fs::OpenOptions;

use std::io;

use std::io::prelude::*;

use std::io::SeekFrom;

use std::path::Path;

Structs

FileReader

Functions

build
Create a new [FileReader] and read/fill with [read_struct]