Struct rbot_discord::core::files::FileReader

pub struct FileReader<T> {
    file: File,
    pub stored: T,
}

Fields
file: File stored: T
Methods

impl<T> FileReader<T> where
T: Deserialize<'de>,
T: Serialize,

pub fn read(&mut self) -> Result<String>

pub fn write_string(&mut self, content: String) -> Result<()>

pub fn write_stored(&mut self) -> Result<()>

pub fn read_struct(&mut self)

Read the file, deserialise it automaticly and store it in the Structure self.stored

pub fn read_json(&mut self) -> Result<Value>

Read the file, deserialise it automaticly and return a Value