Module rbot_discord::features::monitor

Monitor our web services, and the swarm using the Traefik Health informations.

Re-exports

use core::connection::CHANNEL_MAIN;

use reqwest;

use serde_json::from_str;

use serde_json::Value;

use std::thread;

use std::time;

Functions

display_codes

error_code_check
Main loop on a fixed time, get, parse and check for the error codes.

get_codes
Request the codes and parse them, returning a Value

request
Simple get request, returned as string