Module rbot_discord::features::slackimport

Import slack to discord, using exported files

Re-exports

use core::files;

use serenity::model::channel::ChannelType;

use serenity::model::channel::GuildChannel;

use serenity::model::id::ChannelId;

use serenity::model::id::GuildId;

use std::fs;

use std::fs::DirEntry;

use std::io;

use std::path::Path;

Structs

Message
Minimal structure representation from a Slack json export file

Constants

_IMPL_DESERIALIZE_FOR_Message

_IMPL_SERIALIZE_FOR_Message

Statics

GUILD
Fixed guild to export slack files to

Functions

import
Entry point for importing Slack to discord

visit_dirs
Visit every entry in a directory

visit_files
Read and parse the files, output the found messages, to the discord associated channel